+++ packages/nimble-components/src/anchor-tabs/tests/anchor-tabs.spec.ts
@@ -31,7 +31,7 @@ describe('AnchorTabs', () => {
 
     describe('without hrefs', () => {
         async function setup(): Promise<Fixture<AnchorTabs>> {
-            return fixture<AnchorTabs>(
+            return await fixture<AnchorTabs>(
                 html`<nimble-anchor-tabs activeid="tab-two">
                     <nimble-anchor-tab></nimble-anchor-tab>
                     <nimble-anchor-tab id="tab-two"></nimble-anchor-tab>
@@ -159,7 +159,7 @@ describe('AnchorTabs', () => {
 
     describe('with hrefs', () => {
         async function setupWithHrefs(): Promise<Fixture<AnchorTabs>> {
-            return fixture<AnchorTabs>(
+            return await fixture<AnchorTabs>(
                 html`<nimble-anchor-tabs activeid="tab-two">
                     <nimble-anchor-tab href="foo"></nimble-anchor-tab>
                     <nimble-anchor-tab