+++ packages/nimble-components/src/select/tests/select.foundation.spec.ts
@@ -20,7 +20,7 @@ import { template as listOptionTemplate } from '../../list-option/template';
  * Timeout for use in async tets.
  */
 export async function timeout(ms = 0): Promise<void> {
-    return new Promise((resolve, _reject) => {
+    await new Promise((resolve, _reject) => {
         window.setTimeout(() => {
             // eslint-disable-next-line no-void
             resolve(void 0);