+++ packages/nimble-components/src/table-column/duration-text/tests/table-column-duration-text.spec.ts
@@ -30,7 +30,7 @@ describe('TableColumnDurationText', () => {
 
     // prettier-ignore
     async function setup(source: ElementReferences): Promise<Fixture<Table<SimpleTableRecord>>> {
-        return fixture<Table<SimpleTableRecord>>(
+        return await fixture<Table<SimpleTableRecord>>(
             html`<${themeProviderTag} lang="en-US">
                     <${tableTag} ${ref('table')} style="width: 700px">
                         <${tableColumnDurationTextTag} ${ref('column1')} field-name="field" group-index="0">