+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,50 @@
 # Change Log - @ni/spright-components
 
-This log was last generated on Wed, 11 Sep 2024 21:31:24 GMT and should not be manually modified.
+<!-- This log was last generated on Thu, 26 Sep 2024 18:24:50 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 4.1.10
+
+Thu, 26 Sep 2024 18:24:50 GMT
+
+### Patches
+
+- Updates for async eslint rule change ([ni/nimble@7ab95a3](https://github.com/ni/nimble/commit/7ab95a37c760071310a52e0ac233978470168f44))
+- Bump @ni/nimble-components to v32.2.10
+
+## 4.1.9
+
+Thu, 26 Sep 2024 17:20:39 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v32.2.9
+
+## 4.1.8
+
+Thu, 26 Sep 2024 15:22:45 GMT
+
+### Patches
+
+- Dependency updates ([ni/nimble@c17dc0b](https://github.com/ni/nimble/commit/c17dc0b532d107a1866580ed8c8d3ef3e4f61e71))
+- Bump @ni/nimble-components to v32.2.8
+
+## 4.1.7
+
+Wed, 25 Sep 2024 22:17:40 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v32.2.7
+
+## 4.1.6
+
+Wed, 18 Sep 2024 20:24:09 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v32.2.6
 
 ## 4.1.5